# Improve ScanNumericLiteral performance