Should not try/catch, use Try*-else instead.

## Before

### Esprima

|       Method |   N |     Mean |     Error |    StdDev |    Gen 0 | Allocated |
|------------- |---- |---------:|----------:|----------:|---------:|----------:|
| ParseProgram | 200 | 11.59 ms | 0.0685 ms | 0.0572 ms | 953.1250 |   3.84 MB |

## After 

### Esprima

|       Method |   N |     Mean |     Error |    StdDev |    Gen 0 | Allocated |
|------------- |---- |---------:|----------:|----------:|---------:|----------:|
| ParseProgram | 200 | 4.116 ms | 0.0153 ms | 0.0143 ms | 937.5000 |   3.78 MB |

### Jint - EvaluationBenchmark

| Method |  N |     Mean |     Error |    StdDev |    Gen 0 | Allocated |
|------- |--- |---------:|----------:|----------:|---------:|----------:|
|   Jint | 20 | 793.8 us | 0.9599 us | 0.8509 us | 136.7188 | 562.97 KB |